Create Stoer_Wargner.cpp
It is the realization of Simple Stoer_wargner's algorithm. This algorithm is based on the idea of shrinking the graph by merging the most intensive vertices until the graph only contains two combined vertex sets. 
For the MinCutPhase, a single run of it needs at most O(|E|+|V|log |V|)} time which is called |V| -1 times by the MinCut function.
Therefore, the overall running time should be the product of two-phase complexity, which is O(|V||E|+ V log |V|).